Refactor: 포커스 액션 리팩토링
포커스 이전, 다음, 클릭 이동에 대한 액션 리팩토링
