+++ client/src/reducers/CellReducerHandler/index.js
@@ -1,4 +1,5 @@
 import common from "./common";
+import focus from "./focus";
 import block from "./block";
 
-export { common, block };
+export { common, focus, block };
